@@ -1,35 +1,45 @@
-cmake_minimum_required(VERSION 3.1.0)
+cmake_minimum_required(VERSION 3.10...3.22)
 
 list(APPEND CMAKE_MODULE_PATH ${CMAKE_CURRENT_SOURCE_DIR}/cmake)
 include(DefaultBuildType)
 
+# configure vcpkg from environment variables, if present
+include(VcpkgEnv)
+
 # ==== Project Name ====
-project(ouster_example VERSION 20220429)
+project(ouster_example VERSION 20220826)
+
+# generate version header
+set(OusterSDK_VERSION_STRING 0.5.1)
+include(VersionGen)
 
 # ==== Options ====
 option(CMAKE_POSITION_INDEPENDENT_CODE "Build position independent code." ON)
 option(BUILD_SHARED_LIBS "Build shared libraries." OFF)
-option(BUILD_PCAP "Build pcap utils." OFF)
-option(BUILD_TESTING "Build tests" OFF)
+option(BUILD_PCAP "Build pcap utils." ON)
 option(BUILD_VIZ "Build Ouster visualizer." ON)
+option(BUILD_TESTING "Build tests" OFF)
 option(BUILD_EXAMPLES "Build C++ examples" OFF)
+option(OUSTER_USE_EIGEN_MAX_ALIGN_BYTES_32 "Eigen max aligned bytes." ON)
 
-# when building as a top-level project and not as conan library
+# when building as a top-level project
 if(CMAKE_PROJECT_NAME STREQUAL PROJECT_NAME)
-  if(NOT "${CMAKE_CXX_STANDARD}")
-    set(CMAKE_CXX_STANDARD 11)
-  endif()
-  set(CMAKE_CXX_STANDARD_REQUIRED ON)
+  set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
 
-  if(NOT "${CMAKE_CXX_EXTENSIONS}")
+  if(NOT DEFINED CMAKE_CXX_STANDARD)
+    set(CMAKE_CXX_STANDARD 14)
+    set(CMAKE_CXX_STANDARD_REQUIRED ON)
+  endif()
+
+  if(NOT DEFINED CMAKE_CXX_EXTENSIONS)
     set(CMAKE_CXX_EXTENSIONS OFF)
   endif()
 
   if(MSVC)
-    add_compile_options(/W3)
+    add_compile_options(/W2)
     add_compile_definitions(NOMINMAX _USE_MATH_DEFINES WIN32_LEAN_AND_MEAN)
   else()
-    add_compile_options(-Wall -Wextra -Werror -Wno-error=deprecated-declarations)
+    add_compile_options(-Wall -Wextra)
   endif()
 
   include(CTest)
