@@ -66,8 +66,9 @@ static const Table<ChanField, FieldInfo, 8> legacy_field_info{{
     {ChanField::RAW32_WORD3, {UINT32, 8, 0, 0}},
 }};
 
-static const Table<ChanField, FieldInfo, 4> lb_field_info{{
+static const Table<ChanField, FieldInfo, 5> lb_field_info{{
     {ChanField::RANGE, {UINT16, 0, 0x7fff, -3}},
+    {ChanField::FLAGS, {UINT8, 1, 0b10000000, 7}},
     {ChanField::REFLECTIVITY, {UINT8, 2, 0, 0}},
     {ChanField::NEAR_IR, {UINT8, 3, 0, -4}},
     {ChanField::RAW32_WORD1, {UINT32, 0, 0, 0}},
@@ -89,13 +90,24 @@ static const Table<ChanField, FieldInfo, 13> dual_field_info{{
     {ChanField::RAW32_WORD4, {UINT32, 12, 0, 0}},
 }};
 
+static const Table<ChanField, FieldInfo, 8> single_field_info{{
+    {ChanField::RANGE, {UINT32, 0, 0x0007ffff, 0}},
+    {ChanField::FLAGS, {UINT8, 2, 0b11111000, 3}},
+    {ChanField::REFLECTIVITY, {UINT8, 4, 0, 0}},
+    {ChanField::SIGNAL, {UINT16, 6, 0, 0}},
+    {ChanField::NEAR_IR, {UINT16, 8, 0, 0}},
+    {ChanField::RAW32_WORD1, {UINT32, 0, 0, 0}},
+    {ChanField::RAW32_WORD2, {UINT32, 4, 0, 0}},
+    {ChanField::RAW32_WORD3, {UINT32, 8, 0, 0}},
+}};
+
 Table<UDPProfileLidar, ProfileEntry, 32> profiles{{
     {UDPProfileLidar::PROFILE_LIDAR_LEGACY,
      {legacy_field_info.data(), legacy_field_info.size(), 12}},
     {UDPProfileLidar::PROFILE_RNG19_RFL8_SIG16_NIR16_DUAL,
      {dual_field_info.data(), dual_field_info.size(), 16}},
     {UDPProfileLidar::PROFILE_RNG19_RFL8_SIG16_NIR16,
-     {legacy_field_info.data(), legacy_field_info.size(), 12}},
+     {single_field_info.data(), single_field_info.size(), 12}},
     {UDPProfileLidar::PROFILE_RNG15_RFL8_NIR8,
      {lb_field_info.data(), lb_field_info.size(), 4}},
 }};
@@ -104,9 +116,7 @@ static const ProfileEntry& lookup_profile_entry(UDPProfileLidar profile) {
     auto end = profiles.end();
     auto it =
         std::find_if(impl::profiles.begin(), end,
-                     [&](const std::pair<UDPProfileLidar, ProfileEntry>& kv) {
-                         return kv.first == profile;
-                     });
+                     [profile](const auto& kv) { return kv.first == profile; });
 
     if (it == end || it->first == 0)
         throw std::invalid_argument("Unknown lidar udp profile");
@@ -251,6 +261,7 @@ packet_format::FieldIter packet_format::end() const {
 
 uint16_t packet_format::packet_type(const uint8_t* lidar_buf) const {
     if (udp_profile_lidar == UDPProfileLidar::PROFILE_LIDAR_LEGACY) {
+        // LEGACY profile has no packet_type - use 0 to code as 'legacy'
         return 0;
     } else {
         uint16_t res;
@@ -271,6 +282,7 @@ uint16_t packet_format::frame_id(const uint8_t* lidar_buf) const {
 
 uint32_t packet_format::init_id(const uint8_t* lidar_buf) const {
     if (udp_profile_lidar == UDPProfileLidar::PROFILE_LIDAR_LEGACY) {
+        // LEGACY profile has no init_id - use 0 to code as 'legacy'
         return 0;
     } else {
         uint32_t res;
@@ -281,6 +293,8 @@ uint32_t packet_format::init_id(const uint8_t* lidar_buf) const {
 
 uint64_t packet_format::prod_sn(const uint8_t* lidar_buf) const {
     if (udp_profile_lidar == UDPProfileLidar::PROFILE_LIDAR_LEGACY) {
+        // LEGACY profile has no prod_sn (serial number) - use 0 to code as
+        // 'legacy'
         return 0;
     } else {
         uint64_t res;
@@ -298,7 +312,11 @@ const uint8_t* packet_format::nth_col(int n, const uint8_t* lidar_buf) const {
 uint32_t packet_format::col_status(const uint8_t* col_buf) const {
     uint32_t res;
     std::memcpy(&res, col_buf + impl_->status_offset, sizeof(uint32_t));
-    return res;
+    if (udp_profile_lidar == UDPProfileLidar::PROFILE_LIDAR_LEGACY) {
+        return res;  // LEGACY was 32 bits of all 1s
+    } else {
+        return res & 0xffff;  // For eUDP packets, we want the last 16 bits
+    }
 }
 
 uint64_t packet_format::col_timestamp(const uint8_t* col_buf) const {
@@ -436,7 +454,7 @@ const packet_format& get_format(const sensor_info& info) {
 
     std::lock_guard<std::mutex> lk{cache_mx};
     if (!cache.count(k)) {
-        cache[k] = std::unique_ptr<packet_format>(new packet_format{info});
+        cache[k] = std::make_unique<packet_format>(info);
     }
 
     return *cache.at(k);
