@@ -1,12 +1,11 @@
 import os
+import re
 from conans import ConanFile, CMake, tools
 
 from pprint import pformat
 
-
 class OusterSDKConan(ConanFile):
     name = "ouster_sdk"
-    version = "0.4.0"
     license = "BSD 3-Clause License"
     author = "Ouster, Inc."
     url = "https://github.com/ouster-lidar/ouster_example"
@@ -19,14 +18,16 @@ class OusterSDKConan(ConanFile):
         "build_pcap": [True, False],
         "shared": [True, False],
         "fPIC": [True, False],
-        "ensure_cpp17": [True, False]
+        "ensure_cpp17": [True, False],
+        "eigen_max_align_bytes": [True, False],
     }
     default_options = {
         "build_viz": False,
         "build_pcap": False,
         "shared": False,
         "fPIC": True,
-        "ensure_cpp17": False
+        "ensure_cpp17": False,
+        "eigen_max_align_bytes": False,
     }
 
     generators = "cmake_paths", "cmake_find_package"
@@ -44,6 +45,12 @@ class OusterSDKConan(ConanFile):
         "README.rst"
     ]
 
+    # https://docs.conan.io/en/1.51/howtos/capture_version.html#how-to-capture-package-version-from-text-or-build-files
+    def set_version(self):
+        content = tools.load(os.path.join(self.recipe_folder, "CMakeLists.txt"))
+        version = re.search("set\(OusterSDK_VERSION_STRING (.*)\)", content).group(1)
+        self.version = version.strip()
+
     def config_options(self):
         if self.settings.os == "Windows":
             del self.options.fPIC
@@ -51,22 +58,24 @@ class OusterSDKConan(ConanFile):
     def requirements(self):
         self.requires("eigen/3.4.0")
         self.requires("jsoncpp/1.9.5")
+        self.requires("libcurl/7.82.0")
 
         if self.options.build_pcap:
             self.requires("libtins/4.3")
 
         if self.options.build_viz:
-            self.requires("glad/0.1.35")
-            # glew is optional, and probably will not be needed
-            # self.requires("glew/2.2.0")
+            self.requires("glad/0.1.34")
+            if self.settings.os != "Windows":
+                self.requires("xorg/system")
             self.requires("glfw/3.3.6")
             # maybe needed for cpp examples, but not for the lib
             # self.requires("tclap/1.2.4")
 
     def configure_cmake(self):
         cmake = CMake(self)
-        cmake.definitions["BUILD_VIZ"] = True if self.options.build_viz else False
-        cmake.definitions["BUILD_PCAP"] = True if self.options.build_pcap else False
+        cmake.definitions["BUILD_VIZ"] = self.options.build_viz
+        cmake.definitions["BUILD_PCAP"] = self.options.build_pcap
+        cmake.definitions["USE_EIGEN_MAX_ALIGN_BYTES_32"] = self.options.eigen_max_align_bytes
         # alt way, but we use CMAKE_TOOLCHAIN_FILE in other pipeline so avoid overwrite
         # cmake.definitions["CMAKE_TOOLCHAIN_FILE"] = os.path.join(self.build_folder, "conan_paths.cmake")
         cmake.definitions[
@@ -74,7 +83,7 @@ class OusterSDKConan(ConanFile):
                 self.build_folder, "conan_paths.cmake")
         cmake.definitions["BUILD_SHARED_LIBS"] = True if self.options.shared else False
         cmake.definitions["CMAKE_POSITION_INDEPENDENT_CODE"] = (
-            True if self.options.fPIC else False
+            True if "fPIC" in self.options and self.options.fPIC else False
         )
 
         # we use this option until we remove nonstd::optional from SDK codebase (soon)
