@@ -97,7 +97,8 @@ bool next_packet_info(playback_handle& handle, packet_info& info);
 size_t read_packet(playback_handle& handle, uint8_t* buf, size_t buffer_size);
 
 /**
- * Initialize the record handle for recording pcap files.
+ * Initialize the record handle for recording single sensor pcap files. Will be
+ * removed
  *
  * @param[in] file The file path to the target pcap to record to.
  * @param[in] src_ip The source address to label the packets with.
@@ -105,10 +106,21 @@ size_t read_packet(playback_handle& handle, uint8_t* buf, size_t buffer_size);
  * @param[in] frag_size The size of the fragments for packet fragmentation.
  * @param[in] use_sll_encapsulation Whether to use sll encapsulation.
  */
-std::shared_ptr<record_handle> record_initialize(
+[[deprecated]] std::shared_ptr<record_handle> record_initialize(
     const std::string& file, const std::string& src_ip,
     const std::string& dst_ip, int frag_size,
     bool use_sll_encapsulation = false);
+
+/**
+ * Initialize the record handle for recording multi sensor pcap files. Source
+ * and destination IPs must be provided with each packet.
+ *
+ * @param[in] file The file path to the target pcap to record to.
+ * @param[in] frag_size The size of the fragments for packet fragmentation.
+ * @param[in] use_sll_encapsulation Whether to use sll encapsulation.
+ */
+std::shared_ptr<record_handle> record_initialize(
+    const std::string& file, int frag_size, bool use_sll_encapsulation = false);
 
 /**
  * Uninitialize the record handle, closing underlying file.
@@ -118,7 +130,8 @@ std::shared_ptr<record_handle> record_initialize(
 void record_uninitialize(record_handle& handle);
 
 /**
- * Record a buffer to a record_handle pcap file.
+ * Record a buffer to a single sensor record_handle pcap file. Source
+ * and destination IPs must be provided during initialization. Will be removed.
  *
  * @param[in] handle The record handle that record_initialize has initted.
  * @param[in] src_port The source port to label the packets with.
@@ -128,7 +141,26 @@ void record_uninitialize(record_handle& handle);
  * @param[in] microsecond_timestamp The timestamp to record the packet as
  * microseconds.
  */
-void record_packet(record_handle& handle, int src_port, int dst_port,
+[[deprecated]] void record_packet(record_handle& handle, int src_port,
+                                  int dst_port, const uint8_t* buf,
+                                  size_t buffer_size,
+                                  uint64_t microsecond_timestamp);
+
+/**
+ * Record a buffer to a multi sensor record_handle pcap file.
+ *
+ * @param[in] handle The record handle that record_initialize has initted.
+ * @param[in] src_ip The source address to label the packets with.
+ * @param[in] dst_ip The destination address to label the packets with.
+ * @param[in] src_port The source port to label the packets with.
+ * @param[in] dst_port The destination port to label the packets with.
+ * @param[in] buf The buffer to record to the pcap file.
+ * @param[in] buffer_size The size of the buffer to record to the pcap file.
+ * @param[in] microsecond_timestamp The timestamp to record the packet as
+ * microseconds.
+ */
+void record_packet(record_handle& handle, const std::string& src_ip,
+                   const std::string& dst_ip, int src_port, int dst_port,
                    const uint8_t* buf, size_t buffer_size,
                    uint64_t microsecond_timestamp);
 
