@@ -11,6 +11,7 @@
 #include <sstream>
 #include <string>
 
+#include "ouster/impl/build.h"
 #include "ouster/os_pcap.h"
 
 using namespace ouster::sensor_utils;
@@ -84,9 +85,18 @@ This module is generated from the C++ code and not meant to be used directly.
     // pcap writing
     py::class_<std::shared_ptr<record_handle>>(m, "record_handle");
 
-    m.def("record_initialize", &record_initialize, py::arg("file_name"),
-          py::arg("src_ip"), py::arg("dst_ip"), py::arg("frag_size"),
-          py::arg("use_sll_encapsulation") = false);
+    m.def("record_initialize",
+          py::overload_cast<const std::string&, const std::string&,
+                            const std::string&, int, bool>(&record_initialize),
+          py::arg("file_name"), py::arg("src_ip"), py::arg("dst_ip"),
+          py::arg("frag_size"), py::arg("use_sll_encapsulation") = false,
+          R"(
+                ``def record_initialize(file_name: str, src_ip: str, dst_ip: str, frag_size: int,
+                      use_sll_encapsulation: bool = ...) -> record_handle:``
+                  
+                  Initialize record handle for single sensor pcap files
+
+            )");
 
     m.def("record_uninitialize", [](std::shared_ptr<record_handle>& handle) {
         record_uninitialize(*handle);
@@ -104,6 +114,8 @@ This module is generated from the C++ code and not meant to be used directly.
                             static_cast<uint8_t*>(info.ptr), info.size,
                             llround(timestamp * 1e6));
           });
+
+    m.attr("__version__") = ouster::SDK_VERSION;
 
     return m.ptr();
 }
