@@ -19,6 +19,7 @@
 #include <utility>
 #include <vector>
 
+#include "ouster/impl/build.h"
 #include "ouster/lidar_scan.h"
 #include "ouster/point_viz.h"
 #include "ouster/types.h"
@@ -68,7 +69,7 @@ using pymatrixd =
 
 PYBIND11_PLUGIN(_viz) {
     py::module m("_viz", R"(
-    PointViz and LidarScabViz bindings generated by pybind11.
+    PointViz bindings generated by pybind11.
 
     This module is generated from the C++ code and not meant to be used directly.
     )");
@@ -596,7 +597,7 @@ PYBIND11_PLUGIN(_viz) {
         {static_cast<pysize>(viz::calref_n), static_cast<pysize>(3)},
         &viz::calref_palette[0][0]};
 
-    m.attr("__version__") = VERSION_INFO;
+    m.attr("__version__") = ouster::SDK_VERSION;
 
     return m.ptr();
 }
