@@ -115,7 +115,12 @@ def filter_3d_by_range_and_azimuth(hostname: str,
         lidar_port: UDP port to listen on for lidar data
         range_min: range minimum in meters
     """
-    import matplotlib.pyplot as plt  # type: ignore
+    try:
+        import matplotlib.pyplot as plt  # type: ignore
+    except ModuleNotFoundError:
+        print("This example requires matplotlib and an appropriate Matplotlib "
+            "GUI backend such as TkAgg or Qt5Agg.")
+        exit(1)
     import math
 
     # set up figure
@@ -154,8 +159,8 @@ def filter_3d_by_range_and_azimuth(hostname: str,
     plt.show()
 
 
-def live_plot_signal(hostname: str, lidar_port: int = 7502) -> None:
-    """Display signal from live sensor
+def live_plot_reflectivity(hostname: str, lidar_port: int = 7502) -> None:
+    """Display reflectivity from live sensor
 
     Args:
         hostname: hostname of the sensor
@@ -166,7 +171,7 @@ def live_plot_signal(hostname: str, lidar_port: int = 7502) -> None:
 
     print("press ESC from visualization to exit")
 
-    # [doc-stag-live-plot-signal]
+    # [doc-stag-live-plot-reflectivity]
     # establish sensor connection
     with closing(client.Scans.stream(hostname, lidar_port,
                                      complete=False)) as stream:
@@ -175,12 +180,12 @@ def live_plot_signal(hostname: str, lidar_port: int = 7502) -> None:
             for scan in stream:
                 # uncomment if you'd like to see frame id printed
                 # print("frame id: {} ".format(scan.frame_id))
-                signal = client.destagger(stream.metadata,
-                                          scan.field(client.ChanField.SIGNAL))
-                signal = (signal / np.max(signal) * 255).astype(np.uint8)
-                cv2.imshow("scaled signal", signal)
+                reflectivity = client.destagger(stream.metadata,
+                                          scan.field(client.ChanField.REFLECTIVITY))
+                reflectivity = (reflectivity / np.max(reflectivity) * 255).astype(np.uint8)
+                cv2.imshow("scaled reflectivity", reflectivity)
                 key = cv2.waitKey(1) & 0xFF
-                # [doc-etag-live-plot-signal]
+                # [doc-etag-live-plot-reflectivity]
                 # 27 is esc
                 if key == 27:
                     show = False
@@ -274,7 +279,7 @@ def main() -> None:
         "configure-sensor": configure_sensor_params,
         "fetch-metadata": fetch_metadata,
         "filter-3d-by-range-and-azimuth": filter_3d_by_range_and_azimuth,
-        "live-plot-signal": live_plot_signal,
+        "live-plot-reflectivity": live_plot_reflectivity,
         "plot-xyz-points": plot_xyz_points,
         "record-pcap": record_pcap,
     }
