@@ -8,6 +8,7 @@ from copy import copy
 import json
 import numpy
 import pytest
+import inspect
 
 from ouster import client
 
@@ -40,6 +41,7 @@ def test_timestamp_mode_misc() -> None:
     (client.LidarMode.MODE_1024x10, 1024, 10, "1024x10"),
     (client.LidarMode.MODE_1024x20, 1024, 20, "1024x20"),
     (client.LidarMode.MODE_2048x10, 2048, 10, "2048x10"),
+    (client.LidarMode.MODE_4096x5, 4096, 5, "4096x5"),
 ])
 def test_lidar_mode(mode, cols, frequency, string) -> None:
     """Check lidar mode (un)parsing and cols/frequency."""
@@ -53,7 +55,7 @@ def test_lidar_mode(mode, cols, frequency, string) -> None:
 def test_lidar_mode_misc() -> None:
     """Check some misc properties of lidar mode."""
     assert len(
-        client.LidarMode.__members__) == 6, "Don't forget to update tests!"
+        client.LidarMode.__members__) == 7, "Don't forget to update tests!"
     assert client.LidarMode.from_string('foo') == client.LidarMode.MODE_UNSPEC
     assert client.LidarMode(0) == client.LidarMode.MODE_UNSPEC
     assert str(client.LidarMode.MODE_UNSPEC) == "UNKNOWN"
@@ -187,3 +189,21 @@ def test_parse_info() -> None:
     metadata['lidar_origin_to_beam_origin_mm'] = 'foo'
     with pytest.raises(RuntimeError):
         client.SensorInfo(json.dumps(metadata))
+
+
+def test_info_length() -> None:
+    """Check length of info to ensure we've added appropriately to the == operator"""
+
+    info_attributes = inspect.getmembers(client.SensorInfo, lambda a: not inspect.isroutine(a))
+    info_properties = [a for a in info_attributes if not (a[0].startswith('__') and a[0].endswith('__'))]
+
+    assert len(info_properties) == 15, "Don't forget to update tests and the sensor_info == operator!"
+
+
+def test_equality_format() -> None:
+    """Check length of data format to ensure we've added appropriately to the == operator"""
+
+    data_format_attributes = inspect.getmembers(client.DataFormat, lambda a: not inspect.isroutine(a))
+    data_format_properties = [a for a in data_format_attributes if not (a[0].startswith('__') and a[0].endswith('__'))]
+
+    assert len(data_format_properties) == 7, "Don't forget to update tests and the data_format == operator!"
