@@ -52,7 +52,9 @@ testing::Values(
   "2_2_os-992119000444-128_legacy",
   "2_2_os-992119000444-128",
   "2_3_os-992146000760-128_legacy",
-  "2_3_os-992146000760-128"
+  "2_3_os-992146000760-128",
+  "2_3_1_os-992146000760-128_legacy",
+  "2_3_1_os-992146000760-128"
 
 ));
 // clang-format on
